WIP: Enable Auto-Vacuum on CA1 via "Ext. Control"
This commit adds a new configuration setting to enable the "Ext. Control"
switch to function as vacuum control for the CA1 model. enabling auto-vacuum
functionality.

This commit is still a WIP, as currently the external control switch is
activated as soon as the "Auto Vacuum" switch is toggled, and does not
deactivate when the job has finished.